Corrección del scanner

+++ src/_01_PSP_HILOS_PRIMOS/_01_PSP_HILOS_PRIMO.java
@@ -39,7 +39,7 @@ public class _01_PSP_HILOS_PRIMO implements Runnable {
 			return false;
 		}
 
-		for (int i = 5; i * i <= num; i += 6) {
+		for (long i = 5; i * i <= num; i += 6) {
 			if (num % i == 0 || num % (i + 2) == 0) {
 				return false;
 			}
